Email Scrapper/bakend/Emailscrapper/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
dj-database-url==2.2.0
Django==5.1.1
django-background-tasks==1.2.8
django-bootstrap4==24.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
fontawesome-free==5.15.4
idna==3.10
importlib_metadata==8.5.0
Markdown==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
requests==2.32.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
zipp==3.20.2
